cpu/allwinner/a10: Use SD card for bootblock media
This is a brute-force approach of copying the entire CBFS to RAM in
the bootblock. While this may be sub-optimal with large CBFS sizes,
it simplifies the logic of buffer management. We still do not have
a proper CBFS infrastructure for handling this.

With this patch, it is now possible to load a uboot payload, and boot
linux from SD cards.

Change-Id: Ie0b123023c53e485c41650c1fc7b0fb69a286a30
Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>