vendorcode/amd/agesa/fam15tn: Build as a static library
FORTEST. NOTFORMERGEYET.

Up until now, we were building AGESA by specifying each AGESA source
file and adding it to the list of romstage and ramstage source files.
As a result, we were compiling each AGESA source twice, despite the
fact that it does not depend on the stage we're in.

To sweeten the deal, we also remove the practice of specifying every
single AGESA directory as an include dir. We are also able to avoid
polluting coreboot CFLAGS with the AGESA CFLAGS, thus making AGESA an
almost fully self-contained module.

Change-Id: I9b23264129d1c08cb67cabc31d15a68d43ed7624
Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>