arch/x86/Makefile.inc: Re-enable -Wa,--divide assembler flag
In the follow-on patches, we use constant division to calculate the
number of elements in an assembly structure. Since the size of one
element is not a power of two for all structures, we can't bit shift.
This allows us to do the computation at assembly time rather than
runtime.

Change-Id: Ic5693f39dee8daf759bfc16c79ac3c7ddb4de465
Signed-off-by: Alexandru Gagniuc <alexandrux.gagniuc@intel.com>