soc/apollolake: clear whole contents of cache-as-ram
Some code makes assumption on memory to be cleared. In particular
we must clear area occupied by car.data. FSP seems to be also
making assumption stack memory is zeroed. This can be further
reworked to clear only areas that have to be cleared for the sake
of saving cycles.

Change-Id: I9d75cb0a15f2bebf47d53daf7e4d76865a4c166b
Signed-off-by: Andrey Petrov <andrey.petrov@intel.com>