soc/apollolake: Override default mmap_boot.c for all stages
The Apollolake-specific mmap_boot.c implementation was only compiled
for the bootblock. This worked because the generic one was never
included in the bootblock, so we didn't get duplicate symbol.

However, the default mmap_boot won't work on this SOC, so we need to
override it for all stages where it is used.

Change-Id: If633349e0700437d8adfe6af9da045e80c9f02c5
Signed-off-by: Alexandru Gagniuc <alexandrux.gagniuc@intel.com>