drivers/intel/fsp2_0: Return memory_init HOB list to caller
FspMemoryInit creates a list of hand-off-blocks (HOBs) which it
returns to the caller. One of the HOBs describes an area of memory
reserved for FSP. If that area is overwritten, FSP will Crash'n'Burn (TM),
so in a future patch, we will parse the HOB list and mark the FSP
memory area as reserved.

Change-Id: I5ecbb536ac73245aa1b5a433d3dd8ac6eb4b5411
Signed-off-by: Alexandru Gagniuc <alexandrux.gagniuc@intel.com>