soc/apollolake: Tear down cache-as-ram very early in ramstage
Because of the reasons explained in the file, teardown of CAR needs to
be done in ramstage. This patch addresses that issues in a way
designed to minimize boot time. Teardown currently takes about 7 ms,
but that time can be cut to half a millisecond by excluding the fsp
reserved memory range from the flushed ranges.

Change-Id: I6a0cb07f5b35ad42537da688e376ac96612990a6
Signed-off-by: Alexandru Gagniuc <alexandrux.gagniuc@intel.com>