libpayload: x86/timer: Implement constant TSC based calibration
On certain SOCs, such as Apollolake, the 8254 timer is not present or
not operational. On such systems, attempting to calibrate the TSC
against the 8254 will hang the payload. The next best approach is to
identify processors with a constant TSC rate, and determine the actual
speed by reading model-specific registers.

With this patch, filo can now succesfully load the filo shell on
Appololalke RVP2 board.

Change-Id: I863acbc015cf072fa007584ab1d4f4531f0a6d4f
Signed-off-by: Alexandru Gagniuc <alexandrux.gagniuc@intel.com>
Signed-off-by: Andrey Petrov <andrey.petrov@intel.com>