drivers/intel/fsp2_0: Take cbmem overhead size as a parameter
FSP will reserve an area of memory at the top of usable RAM below 4GiB
if asked. Rather than selecting the region's size in fsp_memory_init(),
pass it as a parameter. This allows the caller to handle all the cbmem
related logic, and removes the cbmem dependency from the FSP wrapper.

Change-Id: I581b224ee83202732ae83a0f363f0bb717aae669
Signed-off-by: Alexandru Gagniuc <alexandrux.gagniuc@intel.com>