drivers/intel/fsp2_0: Print names of hand-off-block owner
To help with debugging, also print the name of the owner of a
hand-off-block (HOB), instead of just the UUID. Looking at an actual
string instead of a list of hexadecimal character significantly speeds
up debugging. Only UUIDs which are currently returned by FSP are
added. All others are marked as "UNKNOWN"

Change-Id: I2f3f7579722c8bd440f05c43243f33b3fca63d91
Signed-off-by: Alexandru Gagniuc <alexandrux.gagniuc@intel.com>