cpu/qemu-x86: Enable bootblock console
This brings the qemu C environment bootblock closer to the generic
bootblock in lib/. This is a step before the bootblocks can be
converged, but the code is not yet ready for unification.
When running qemu with the "-serial stdio" flag, the
"bootblock starting..." message now appears on the console.

Change-Id: I121bce620cf7f8a713e7dad1833cff1b0dc2116f
Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>