cpu/qemu-x86: Use run_romstage instead of custom inline assembly
Now that passing the BIST result to romstage is no longer required,
we no longer need to use custom assembly to call into romstage.
Instead, a lot of the logic in the bootblock can now be replaced by
a single call to run_romstage(). This is another step in converging
the bootblock with the generic bootblock logic in lib/.

Change-Id: I66436c8fba5121a3758d02955f691ac9d273b049
Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>